Consolidate all education validations (#4445)
* Initial general validations consolidated

* Add consolidation of field validations

* Fix failing unit tests

* Make review fixes
