+++ src/js/edu-benefits/components/military-history/MilitaryServiceTour.jsx
@@ -6,8 +6,8 @@ import ErrorableTextarea from '../../../common/components/form-elements/Errorabl
 import ErrorableDate from '../../../common/components/form-elements/ErrorableDate';
 import ErrorableCurrentOrPastDate from '../../../common/components/form-elements/ErrorableCurrentOrPastDate';
 import ExpandingGroup from '../../../common/components/form-elements/ExpandingGroup';
+import { validateIfDirty, isNotBlank, isValidDateRange } from '../../../common/utils/validations';
 
-import { validateIfDirty, isNotBlank, isValidDateRange } from '../../utils/validations';
 import ServicePeriodsReview from './ServicePeriodsReview';
 
 export default class MilitaryServiceTour extends React.Component {
