+++ src/js/edu-benefits/components/military-history/RotcScholarship.jsx
@@ -2,8 +2,9 @@ import React from 'react';
 
 import ErrorableTextInput from '../../../common/components/form-elements/ErrorableTextInput';
 import ErrorableNumberInput from '../../../common/components/form-elements/ErrorableNumberInput';
+import { validateIfDirty, isValidYear, isValidValue } from '../../../common/utils/validations';
 
-import { isValidYear, validateIfDirty, isValidMonetaryValue, isValidValue } from '../../utils/validations';
+import { isValidMonetaryValue } from '../../utils/validations';
 
 export default class RotcScholarship extends React.Component {
   render() {
