+++ test/edu-benefits/utils/validations.unit.spec.js
@@ -4,12 +4,10 @@ import moment from 'moment';
 import {
   isValidContactInformationPage,
   isValidMilitaryServicePage,
-  isValidDateRange,
   isValidFutureOrPastDateField,
   isValidPage,
   isValidRoutingNumber,
   isValidRelinquishedDate,
-  isValidPartialMonthYearRange,
   isValidEducationPeriod,
   isValidEducationHistoryPage
 } from '../../../src/js/edu-benefits/utils/validations.js';
@@ -18,101 +16,6 @@ import { createVeteran } from '../../../src/js/edu-benefits/utils/veteran.js';
 import { makeField } from '../../../src/js/common/model/fields.js';
 
 describe('Validation:', () => {
-  describe('isValidDateRange', () => {
-    it('validates if to date is after from date', () => {
-      const fromDate = {
-        day: {
-          value: 3,
-          dirty: true
-        },
-        month: {
-          value: 3,
-          dirty: true
-        },
-        year: {
-          value: 2006,
-          dirty: true
-        }
-      };
-      const toDate = {
-        day: {
-          value: 3,
-          dirty: true
-        },
-        month: {
-          value: 4,
-          dirty: true
-        },
-        year: {
-          value: 2006,
-          dirty: true
-        }
-      };
-      expect(isValidDateRange(fromDate, toDate)).to.be.true;
-    });
-    it('does not validate to date is before from date', () => {
-      const fromDate = {
-        day: {
-          value: 3,
-          dirty: true
-        },
-        month: {
-          value: 3,
-          dirty: true
-        },
-        year: {
-          value: 2006,
-          dirty: true
-        }
-      };
-      const toDate = {
-        day: {
-          value: 3,
-          dirty: true
-        },
-        month: {
-          value: 4,
-          dirty: true
-        },
-        year: {
-          value: 2005,
-          dirty: true
-        }
-      };
-      expect(isValidDateRange(fromDate, toDate)).to.be.false;
-    });
-    it('does validate with partial dates', () => {
-      const fromDate = {
-        day: {
-          value: '3',
-          dirty: true
-        },
-        month: {
-          value: 3,
-          dirty: true
-        },
-        year: {
-          value: 2006,
-          dirty: true
-        }
-      };
-      const toDate = {
-        day: {
-          value: '',
-          dirty: true
-        },
-        month: {
-          value: '',
-          dirty: true
-        },
-        year: {
-          value: 2008,
-          dirty: true
-        }
-      };
-      expect(isValidDateRange(fromDate, toDate)).to.be.true;
-    });
-  });
   describe('isValidPage:', () => {
     describe('EmploymentHistory', () => {
       it('validates page without license is valid', () => {
@@ -326,92 +229,6 @@ describe('Validation:', () => {
       };
 
       expect(isValidMilitaryServicePage(data)).to.be.true;
-    });
-  });
-  describe('isValidPartialMonthYearRange', () => {
-    it('should validate partial range', () => {
-      const fromDate = {
-        month: {
-          value: '2'
-        },
-        year: {
-          value: '2001'
-        }
-      };
-
-      const toDate = {
-        month: {
-          value: '3'
-        },
-        year: {
-          value: ''
-        }
-      };
-
-      expect(isValidPartialMonthYearRange(fromDate, toDate)).to.be.true;
-    });
-    it('should not validate invalid range', () => {
-      const fromDate = {
-        month: {
-          value: '2'
-        },
-        year: {
-          value: '2002'
-        }
-      };
-
-      const toDate = {
-        month: {
-          value: '3'
-        },
-        year: {
-          value: '2001'
-        }
-      };
-
-      expect(isValidPartialMonthYearRange(fromDate, toDate)).to.be.false;
-    });
-    it('should validate same date range', () => {
-      const fromDate = {
-        month: {
-          value: '2'
-        },
-        year: {
-          value: '2001'
-        }
-      };
-
-      const toDate = {
-        month: {
-          value: '2'
-        },
-        year: {
-          value: '2001'
-        }
-      };
-
-      expect(isValidPartialMonthYearRange(fromDate, toDate)).to.be.true;
-    });
-    it('should validate year only range', () => {
-      const fromDate = {
-        month: {
-          value: ''
-        },
-        year: {
-          value: '2001'
-        }
-      };
-
-      const toDate = {
-        month: {
-          value: ''
-        },
-        year: {
-          value: '2002'
-        }
-      };
-
-      expect(isValidPartialMonthYearRange(fromDate, toDate)).to.be.true;
     });
   });
   describe('isValidEducationPeriod', () => {
